fix(reports): address review feedback on memory pattern callout
- Replace all em dashes with hyphen-with-spaces (' - ')
- Move MemoryPatternCallout to its own co-located file
- Anchor callout after ram-usage chart instead of swap-usage
- Standardize terminology to 'Cache + Buffers' throughout callout body
